ci: Check all crate documentation in addition to target-specific ndk (#129)
* ci: Check all crate documentation in addition to target-specific ndk

Only `ndk` docs are built for all `android` targets but the rest of the
documentation remains untested. Especially now that we're expanding to
provide more complete docs for ndk-build and cargo-apk should everything
be tested, to make sure that things like intra-doc links remain in
working order.

* ci: Don't build target-agnostic ndk docs for all targets

The previous commit tests all documentation for all features and crates
(not just `ndk`). The NDK crate and its documentation build without
issues thanks to the guards introduced by the `"test"` feature. Since
the `ndk` docs do not depend on any target-specific flags (and `ndk-sys`
does not contain any documentation) is is fine to build-test it just
once for the host architecture.