ndk-macro: Reexport android_logger from ndk-glue and use it in macros (#88)
* ndk-examples: Remove empty lib.rs in favour of examples

This crate does not contain executables nor libraries, just the examples
in the example directory.  This is already confusing [1] [2] as they can
only be run with something like:

    cargo apk run -p ndk-examples --example hello_world

Or cd-ing into `ndk-examples` before being able to run these.

Unfortunately `cargo apk run -p ndk-examples` keeps working fine,
resulting in:

    Error: Path "target/aarch64-linux-android/debug/libndk_examples.so" doesn't exist.

Despite the crate clearly not exposing a library target.  The same
happens without this change because `cdylib` is not set for `[lib]`,
but time is still spent compiling all the dependencies instead of
erroring out straight away.

[1]: https://github.com/rust-windowing/android-ndk-rs/issues/48#issuecomment-656931548
[2]: https://github.com/rust-windowing/android-ndk-rs/issues/29#issuecomment-635395518

* ndk-examples: Use logger feature to print "hello world"

It is more common to use a logging framework instead of printing raw
text to stdout that ends up in the logs; lead by example.

More importantly this exemplifies a current issue with android_logger
not being exported by the relevant macro crate.

* ndk-macro: Reexport android_logger from ndk-glue and use it in macros

When enabling the logger feature:

    error[E0433]: failed to resolve: use of undeclared type or module `android_logger`
     --> ndk-examples/examples/hello_world.rs:6:5
      |
    6 |     ndk_glue::main(backtrace = "on", logger(level = "debug", tag = "hello-world"))
      |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ use of undeclared type or module `android_logger`
      |
      = note: this error originates in an attribute macro (in Nightly builds, run with -Z macro-backtrace for more info)

This means the crate user is responsible for importing the
android_logger crate in their namespace where this macro is used.  That
is not ideal, especially when ndk-glue already imports it but doesn't
seem to use it anywhere.

* ndk-macro: Reexport log crate from ndk-glue and use it in macros

In accordance with a prior, identical change for android_logger:
reexport the log crate from ndk-glue and use it in the macro instead of
relying on crate users to provide log in this scope or override the
path.
This is especially useful if a crate wishes to configure the log level
or tag without needing the log crate in the scope where ndk_glue::main
is called.

* CHANGELOG: Document removal of `android_logger`/`log` override setting