ndk-sys: Switch remaining enums to newtypes with associated constants (#315)
Searching for `pub type` reveals a lot of typedefs for enums that could
have been a newtype with associated constants.  Unfortunately there are
still a lot of private enums in the source, some with a `typedef`
following: these cannot be converted yet, otherwise we might have gotten
away with globally enabling `--default-enum-style` to apply this style
to _every_ enum.

At the same time certain enums could use `--bitfield-enum` yet those
eligible also suffer from the same lack of a type, requiring the user to
work with `_bindgen_ty_XX` which is far from ergonomic.